run-core: account for exec tests with an argument
We prepare the template in a way that we only have to exchange the
argument. In case there is none we remove the whole line as otherwise
we might have to deal with properly quoting '<arg ...' because the
bash could interpret it as a request to read the file called 'arg'.